% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11janitor-package.R
\docType{package}
\name{cpp11janitor-package}
\alias{cpp11janitor}
\alias{cpp11janitor-package}
\title{cpp11janitor: C++ Tools for Cleaning Strings}
\description{
Provides function declarations and inline function definitions that facilitate cleaning strings in C++ code before passing them to R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cpp11janitor/}
  \item \url{https://github.com/pachadotdev/cpp11janitor}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11janitor/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item Munk School of Global Affairs and Public Policy [funder]
}

}
\keyword{internal}
